Bondone Dataset

It contains hourly meteorological data observed at MeteoTrentino T0327 station located at Monte Bondone-Viotte (Trentino, Easter Alps, Italy) from August 2004 to December 2012.\

The zoo object 'meteo' contains:

Iprec

Hourly Precipitation Depth expressed in millimeters

AirT

Air Temperature expressed in Celsius Degree

RH

Relative Humidity in PerCent

WinDir

Wind Direction expressed in Degrees North Clockwise

WinSp

Wind Direction expressed in meters per second

Swglob

Short-Wave Radiation expressed in Watts per square meters

The corresponding time axis vector for each observation can be printed by typing index(meteo).

data(bondone)

Details

This data set stores all meteorological information useful for a GEOtop simulation. The user can easily use the package with his/her own data after replacing the values of such variables.

Format

Data frame , 'zoo' object

Source

Original data are provided by Provincia Autonoma di Trento (https://www.meteotrentino.it/)).

This dataset is intended for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.