Collects all keywords contained in the 'getop.inpts' configuration files and their values in a data frame object.

Collects all keywords contained in the 'getop.inpts' configuration files and their values in a data frame object.

declared.geotop.inpts.keywords(
  wpath,
  inpts.file = "geotop.inpts",
  comment = "!",
  exceptions = "Date",
  warn = FALSE,
  no.comment = c("!>!", "!>>!"),
  ...
)

Arguments

wpath

working directory containing GEOtop files

inpts.file

name of the GEOtop configuration file. Default is "geotop.inpts"

comment

comment indicator charcater. Default is "!"

exceptions

string vector. If keywords contain an element of this vector, the blank spaces in Value " " will not be removed.

warn

logical argument of readLines. Default is FALSE.

no.comment

string indicatos read as comment ones by GEOtop but they do not indicate comments by "geotopbricks" package.

...

further arguments of readLines

Value

a data frame with two columns: Keyword and Value

See also