This function saves all spatially distrubuted information contained in the recovery folder into a comprehensive list object.

This function saves all spatially distrubuted information contained in the recovery folder into a comprehensive list object.

get.geotop.recovery.state(
  recFolder,
  xx = "0000",
  formatter = "L%04d",
  extension = ".asc",
  nsoillayers = 10,
  layersFromDir = FALSE,
  ...
)

Arguments

recFolder

directory when recvery maps are set. In GEOtop it is ...

xx

charcter String. Default is "0000"

formatter

string character for the the decimal formatter to be used. Default is "L%04d".

extension

file estension used for asccii recovery map files. It must contains '.' as the first character. Defaut is ".asc" .

nsoillayers

number of soil layers used in the GEOtop simulation.

layersFromDir

logical value. If is TRUE the number of soil/snow (vertical) layers used in the GEOtop simulation is automatically calculated and cannot be assigned through nsoillayers.

...

further arguments

Value

a list object containining all recovery raster maps.

Note

This function has been used with the built 1.225-9 of GEOtop .

See also

Author

Emanuele Cordano

Examples

library(geotopbricks)
example_Rscript <- system.file('template/example.geotop.recovery.state.R',package="geotopbricks")
example_Rscript
#> [1] "/tmp/RtmpDxnCqu/temp_libpath1eed372e00c4ea/geotopbricks/template/example.geotop.recovery.state.R"

# Not Run because it elapses too long time!!! 
# Please Uncomment the following line to run by yourself!!!
# source(example_Rscript)