pointer.to.maps.xyz.time

'pointer.to.maps.xyz.time' function (obsolete)

pointer.to.maps.xyz.time(
  wpath,
  map.prefix = "thetaliq",
  suffix = "L%04dN%04d.asc",
  zoo.index = NULL,
  ntime,
  nlayers
)

Arguments

wpath

complete working path to *.asc maps are saved

map.prefix

string prefix name map before

suffix

z-time or time suffix plus file extention character string. Default for GEOtop application is "L%04dN%04d.asc" for xy+z+time maps or "N%04d.asc" for xy+time maps.

zoo.index

time or date index. Default is NULL , otherwise function returns a zoo object with zoo.index as index.

ntime

number of time instant. If zoo.index is not NULL, it is calculated from zoo.index length.

nlayers

number of vertical layers.

Value

A data.frame or zoo object containig the paths to maps fpr each time and z layer.

Author

Emanuele Cordano